LayoutTests/imported/w3c/web-platform-tests/html/semantics/embedded-content/media-elements/event_pause_noautoplay-expected.txt

CONSOLE MESSAGE: Unhandled Promise Rejection: AbortError: The operation was aborted.
CONSOLE MESSAGE: Unhandled Promise Rejection: AbortError: The operation was aborted.
spec reference

  

Harness Error (FAIL), message = The operation was aborted.

PASS audio events - pause 
PASS calling play() then pause() on non-autoplay audio should trigger pause event 
PASS video events - pause 
PASS calling play() then pause() on non-autoplay video should trigger pause event 
